LPS-102711 Uses the new contributor class
