+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/portal/settings/configuration/admin/display/CategorizationSiteSettingsConfigurationScreenContributor.java
@@ -14,22 +14,16 @@
 
 package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
 
-import com.liferay.configuration.admin.display.ConfigurationScreen;
-import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-
-import java.io.IOException;
+import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 
 import java.util.Locale;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -37,28 +31,38 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Eudaldo Alonso
  */
-@Component(service = ConfigurationScreen.class)
-public class DefaultUserAssociationsSiteSettingsConfigurationScreen
-	implements ConfigurationScreen {
+@Component(service = SiteSettingsConfigurationScreenContributor.class)
+public class CategorizationSiteSettingsConfigurationScreenContributor
+	implements SiteSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "users";
+		return "assets";
+	}
+
+	@Override
+	public String getJspPath() {
+		return "/site_settings/categorization.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "site-configuration-default-user-associations";
+		return "site-configuration-categorization";
 	}
 
 	@Override
 	public String getName(Locale locale) {
-		return LanguageUtil.get(locale, "default-user-associations");
+		return LanguageUtil.get(locale, "categorization");
 	}
 
 	@Override
-	public String getScope() {
-		return ExtendedObjectClassDefinition.Scope.GROUP.getValue();
+	public String getSaveMVCActionCommandName() {
+		return "/site_admin/edit_categorization";
+	}
+
+	@Override
+	public ServletContext getServletContext() {
+		return _servletContext;
 	}
 
 	@Override
@@ -70,30 +74,11 @@ public class DefaultUserAssociationsSiteSettingsConfigurationScreen
 
 		Group siteGroup = themeDisplay.getSiteGroup();
 
-		if ((siteGroup == null) || siteGroup.isCompany()) {
+		if ((siteGroup != null) && siteGroup.isCompany()) {
 			return false;
 		}
 
 		return true;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		try {
-			RequestDispatcher requestDispatcher =
-				_servletContext.getRequestDispatcher(
-					"/site_settings/default_user_associations.jsp");
-
-			requestDispatcher.include(httpServletRequest, httpServletResponse);
-		}
-		catch (Exception exception) {
-			throw new IOException(
-				"Unable to render default_user_associations.jsp", exception);
-		}
 	}
 
 	@Reference(target = "(osgi.web.symbolicname=com.liferay.site.admin.web)")
