+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/portal/settings/configuration/admin/display/LanguagesSiteSettingsConfigurationScreenContributor.java
@@ -14,18 +14,16 @@
 
 package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
 
-import com.liferay.configuration.admin.display.ConfigurationScreen;
-import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
 import com.liferay.portal.kernel.language.LanguageUtil;
-
-import java.io.IOException;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 
 import java.util.Locale;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -33,46 +31,54 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Eudaldo Alonso
  */
-@Component(service = ConfigurationScreen.class)
-public class DetailsSiteSettingsConfigurationScreen
-	implements ConfigurationScreen {
+@Component(service = SiteSettingsConfigurationScreenContributor.class)
+public class LanguagesSiteSettingsConfigurationScreenContributor
+	implements SiteSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "general";
+		return "localization";
+	}
+
+	@Override
+	public String getJspPath() {
+		return "/site_settings/languages.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "site-configuration-details";
+		return "site-configuration-languages";
 	}
 
 	@Override
 	public String getName(Locale locale) {
-		return LanguageUtil.get(locale, "details");
+		return LanguageUtil.get(locale, "languages");
 	}
 
 	@Override
-	public String getScope() {
-		return ExtendedObjectClassDefinition.Scope.GROUP.getValue();
+	public String getSaveMVCActionCommandName() {
+		return "/site_admin/edit_languages";
 	}
 
 	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
+	public ServletContext getServletContext() {
+		return _servletContext;
+	}
 
-		try {
-			RequestDispatcher requestDispatcher =
-				_servletContext.getRequestDispatcher(
-					"/site_settings/details.jsp");
+	@Override
+	public boolean isVisible() {
+		ServiceContext serviceContext =
+			ServiceContextThreadLocal.getServiceContext();
 
-			requestDispatcher.include(httpServletRequest, httpServletResponse);
+		ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
+
+		Group siteGroup = themeDisplay.getSiteGroup();
+
+		if ((siteGroup == null) || siteGroup.isCompany()) {
+			return false;
 		}
-		catch (Exception exception) {
-			throw new IOException("Unable to render details.jsp", exception);
-		}
+
+		return true;
 	}
 
 	@Reference(target = "(osgi.web.symbolicname=com.liferay.site.admin.web)")
