--- modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/portlet/action/EditMapsMVCActionCommand.java
@@ -1,97 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.site.admin.web.internal.portlet.action;
-
-import com.liferay.configuration.admin.constants.ConfigurationAdminPortletKeys;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.portlet.bridges.mvc.BaseTransactionalMVCActionCommand;
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.GroupService;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextFactory;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.PropertiesParamUtil;
-import com.liferay.portal.kernel.util.UnicodeProperties;
-
-import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Eudaldo Alonso
- */
-@Component(
-	immediate = true,
-	property = {
-		"javax.portlet.name=" + ConfigurationAdminPortletKeys.SITE_SETTINGS,
-		"mvc.command.name=/site_admin/edit_maps"
-	},
-	service = MVCActionCommand.class
-)
-public class EditMapsMVCActionCommand
-	extends BaseTransactionalMVCActionCommand {
-
-	@Override
-	protected void doTransactionalCommand(
-			ActionRequest actionRequest, ActionResponse actionResponse)
-		throws Exception {
-
-		long liveGroupId = ParamUtil.getLong(actionRequest, "liveGroupId");
-
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			Group.class.getName(), actionRequest);
-
-		ServiceContextThreadLocal.pushServiceContext(serviceContext);
-
-		Group liveGroup = _groupLocalService.getGroup(liveGroupId);
-
-		UnicodeProperties typeSettingsUnicodeProperties =
-			liveGroup.getTypeSettingsProperties();
-
-		UnicodeProperties formTypeSettingsUnicodeProperties =
-			PropertiesParamUtil.getProperties(
-				actionRequest, "TypeSettingsProperties--");
-
-		typeSettingsUnicodeProperties.putAll(formTypeSettingsUnicodeProperties);
-
-		if (liveGroup.hasStagingGroup()) {
-			Group stagingGroup = liveGroup.getStagingGroup();
-
-			UnicodeProperties stagedGroupTypeSettingsUnicodeProperties =
-				stagingGroup.getTypeSettingsProperties();
-
-			stagedGroupTypeSettingsUnicodeProperties.putAll(
-				formTypeSettingsUnicodeProperties);
-
-			_groupService.updateGroup(
-				stagingGroup.getGroupId(),
-				stagedGroupTypeSettingsUnicodeProperties.toString());
-		}
-
-		_groupService.updateGroup(
-			liveGroup.getGroupId(), typeSettingsUnicodeProperties.toString());
-	}
-
-	@Reference
-	private GroupLocalService _groupLocalService;
-
-	@Reference
-	private GroupService _groupService;
-
-}
