+++ modules/apps/site/site-admin-web/src/main/resources/META-INF/resources/site_settings/custom_fields.jsp
@@ -29,31 +29,9 @@ else {
 }
 %>
 
-<portlet:actionURL name="/site_admin/edit_custom_fields" var="editCustomFieldsURL">
-	<portlet:param name="mvcRenderCommandName" value="/configuration_admin/view_configuration_screen" />
-	<portlet:param name="configurationScreenKey" value="site-configuration-custom-fields" />
-</portlet:actionURL>
-
-<liferay-frontend:edit-form
-	action="<%= editCustomFieldsURL %>"
-	method="post"
-	name="fm"
->
-	<aui:input name="redirect" type="hidden" value="<%= currentURL %>" />
-	<aui:input name="liveGroupId" type="hidden" value="<%= liveGroup.getGroupId() %>" />
-
-	<liferay-frontend:edit-form-body>
-		<liferay-expando:custom-attribute-list
-			className="<%= Group.class.getName() %>"
-			classPK="<%= liveGroup.getGroupId() %>"
-			editable="<%= true %>"
-			label="<%= true %>"
-		/>
-	</liferay-frontend:edit-form-body>
-
-	<liferay-frontend:edit-form-footer>
-		<aui:button type="submit" />
-
-		<aui:button href='<%= ParamUtil.getString(request, "redirect") %>' type="cancel" />
-	</liferay-frontend:edit-form-footer>
-</liferay-frontend:edit-form>
+<liferay-expando:custom-attribute-list
+	className="<%= Group.class.getName() %>"
+	classPK="<%= liveGroup.getGroupId() %>"
+	editable="<%= true %>"
+	label="<%= true %>"
+/>
