+++ modules/apps/site/site-admin-web/src/main/resources/META-INF/resources/site_settings/languages.jsp
@@ -20,35 +20,13 @@
 DisplaySettingsDisplayContext displaySettingsDisplayContext = new DisplaySettingsDisplayContext(request, liferayPortletResponse);
 %>
 
-<portlet:actionURL name="/site_admin/edit_languages" var="editLanguagesURL">
-	<portlet:param name="mvcRenderCommandName" value="/configuration_admin/view_configuration_screen" />
-	<portlet:param name="configurationScreenKey" value="site-configuration-languages" />
-</portlet:actionURL>
+<div id="<portlet:namespace />siteLanguageConfiguration">
+	<div class="inline-item my-5 p-5 w-100">
+		<span aria-hidden="true" class="loading-animation"></span>
+	</div>
 
-<liferay-frontend:edit-form
-	action="<%= editLanguagesURL %>"
-	method="post"
-	name="fm"
->
-	<aui:input name="redirect" type="hidden" value="<%= currentURL %>" />
-	<aui:input name="liveGroupId" type="hidden" value="<%= displaySettingsDisplayContext.getLiveGroupId() %>" />
-
-	<liferay-frontend:edit-form-body>
-		<div id="<portlet:namespace />siteLanguageConfiguration">
-			<div class="inline-item my-5 p-5 w-100">
-				<span aria-hidden="true" class="loading-animation"></span>
-			</div>
-
-			<react:component
-				module="js/SiteLanguageConfiguration"
-				props="<%= displaySettingsDisplayContext.getPropsMap() %>"
-			/>
-		</div>
-	</liferay-frontend:edit-form-body>
-
-	<liferay-frontend:edit-form-footer>
-		<aui:button type="submit" />
-
-		<aui:button href='<%= ParamUtil.getString(request, "redirect") %>' type="cancel" />
-	</liferay-frontend:edit-form-footer>
-</liferay-frontend:edit-form>
+	<react:component
+		module="js/SiteLanguageConfiguration"
+		props="<%= displaySettingsDisplayContext.getPropsMap() %>"
+	/>
+</div>
